Skipio-bot/requirements.txt
discord
flask
discord.py
cloudscraper
pandas
requests
PyGithub
python-dotenv
psycopg2-binary
supabase
matplotlib
seaborn
scipy
